refactor: Ranking 컨트롤러 분리
