+++ back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/service/PostService.java
@@ -1,11 +1,15 @@
 package wooteco.team.ittabi.legenoaroundhere.service;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,6 +18,7 @@ import wooteco.team.ittabi.legenoaroundhere.config.IAuthenticationFacade;
 import wooteco.team.ittabi.legenoaroundhere.domain.PostSearch;
 import wooteco.team.ittabi.legenoaroundhere.domain.area.Area;
 import wooteco.team.ittabi.legenoaroundhere.domain.post.Post;
+import wooteco.team.ittabi.legenoaroundhere.domain.post.RankingCriteria;
 import wooteco.team.ittabi.legenoaroundhere.domain.post.image.PostImage;
 import wooteco.team.ittabi.legenoaroundhere.domain.sector.Sector;
 import wooteco.team.ittabi.legenoaroundhere.domain.user.User;
@@ -23,6 +28,7 @@ import wooteco.team.ittabi.legenoaroundhere.dto.PostResponse;
 import wooteco.team.ittabi.legenoaroundhere.dto.PostSearchRequest;
 import wooteco.team.ittabi.legenoaroundhere.dto.PostUpdateRequest;
 import wooteco.team.ittabi.legenoaroundhere.dto.PostWithCommentsCountResponse;
+import wooteco.team.ittabi.legenoaroundhere.dto.RankingRequest;
 import wooteco.team.ittabi.legenoaroundhere.exception.NotAuthorizedException;
 import wooteco.team.ittabi.legenoaroundhere.exception.NotExistsException;
 import wooteco.team.ittabi.legenoaroundhere.exception.WrongUserInputException;
@@ -176,5 +182,29 @@ public class PostService {
 
         Post post = findPostBy(postId);
         post.pressZzang(user);
+    }
+
+    @Transactional
+    public Page<PostWithCommentsCountResponse> searchRanking(Pageable pageable,
+        RankingRequest rankingRequest, PostSearchRequest postSearchFilter) {
+        Page<Post> allDateZzangPosts = getPostByFilter(pageable, postSearchFilter.toPostSearch());
+        List<Post> filteredDateZzangPosts
+            = getPostsThatZzangFilteredByDate(rankingRequest.getCriteria(), allDateZzangPosts);
+        Collections.reverse(filteredDateZzangPosts);
+
+        User user = (User) authenticationFacade.getPrincipal();
+        PageImpl<Post> posts = new PageImpl<>(filteredDateZzangPosts);
+        return posts.map(post -> PostWithCommentsCountResponse.of(user, post));
+    }
+
+    private List<Post> getPostsThatZzangFilteredByDate(String criteria,
+        Page<Post> unfilteredPosts) {
+        RankingCriteria rankingCriteria = RankingCriteria.of(criteria);
+        LocalDateTime startDate = rankingCriteria.getStartDate();
+        LocalDateTime endDate = rankingCriteria.getEndDate();
+
+        return unfilteredPosts.stream()
+            .sorted(Comparator.comparing(post -> post.getPostZzangCountByDate(startDate, endDate)))
+            .collect(Collectors.toList());
     }
 }
